Consolidate modules into a single module.
